# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rosshield API
  version: 0.0.1
  summary: ROS2 Fleet Security Audit Platform API
  description: |
    Skeleton spec for Phase 0. Contract-first — server handlers, client SDKs,
    and CLI command shapes are generated from this document. See
    `docs/design/05-api-and-auth.md` for the full design.

    Status: v1 under design. No operation here is considered stable.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local desktop / development
  - url: https://{host}
    description: On-prem or appliance
    variables:
      host:
        default: rosshield.example.internal
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/webhooks:
    get:
      tags: [webhook]
      operationId: listWebhookEndpoints
      summary: List webhook endpoints for the current tenant
      x-status: phase3
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWebhookEndpointsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [webhook]
      operationId: createWebhookEndpoint
      summary: Register a new webhook endpoint
      description: |
        Validates URL (absolute http/https), secret (non-empty for HMAC), and event
        types (must be in known set: scan.completed, insight.created, audit.checkpoint).
      x-status: phase3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/webhooks/{endpointId}:
    parameters:
      - { in: path, name: endpointId, required: true, schema: { type: string, pattern: "^wh_[A-Z0-9]+$" } }
    get:
      tags: [webhook]
      operationId: getWebhookEndpoint
      summary: Get a webhook endpoint
      x-status: phase3
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags: [webhook]
      operationId: updateWebhookEndpoint
      summary: Update a webhook endpoint (URL · secret · events · format · enabled)
      x-status: phase3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [webhook]
      operationId: deleteWebhookEndpoint
      summary: Delete a webhook endpoint (deliveries are preserved — append-only)
      x-status: phase3
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `/api/v1/auth/login` or OIDC/SAML callback
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API key in `Authorization: Bearer fg_live_...` form"
...
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:     { type: string }
        serverVersion: { type: string }
        responseAt:    { type: string, format: date-time }
...
    ErrorCategory:
      type: string
      enum:
        - unauthenticated
        - forbidden
        - not_found
        - conflict
        - validation
        - rate_limited
        - upstream
        - internal
    Error:
      type: object
      required: [code, category, message]
      properties:
        code:     { type: string, description: "Dotted code (e.g. `robot.not_found`)" }
        category: { $ref: "#/components/schemas/ErrorCategory" }
        message:  { type: string }
        target:
          type: object
          properties:
            type: { type: string }
            id:   { type: string }
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:    { type: boolean, const: false }
        error: { $ref: "#/components/schemas/Error" }
        meta:  { $ref: "#/components/schemas/Meta" }

    # ─── Robot ────────────────────────────────────────────────────────────
...
    WebhookEventType:
      type: string
      description: |
        Domain event type subscribed by the webhook. Empty `events` means all known
        events. Add new types only after operator agreement (signed pack).
      enum: [scan.completed, insight.created, audit.checkpoint]
    WebhookFormat:
      type: string
      description: |
        Payload serialization format.
          - json (default): rosshield raw JSON envelope.
          - cef: ArcSight Common Event Format (Splunk OOTB).
          - ecs: Elastic Common Schema.
      enum: [json, cef, ecs]
    WebhookEndpoint:
      type: object
      required: [id, tenantId, url, secretLast4, events, format, enabled, createdAt, updatedAt]
      properties:
        id:          { type: string, pattern: "^wh_[A-Z0-9]+$" }
        tenantId:    { type: string }
        url:         { type: string, format: uri, description: "Absolute http/https URL" }
        secretLast4:
          type: string
          description: |
            Last 4 chars of the HMAC-SHA256 signing secret (`****` if shorter than 4).
            Full secret is never returned — operator must manage rotation via the
            UPDATE endpoint or external secret manager.
        events:
          type: array
          description: "Empty array → subscribes to all known event types."
          items: { $ref: "#/components/schemas/WebhookEventType" }
        format:      { $ref: "#/components/schemas/WebhookFormat" }
        enabled:     { type: boolean }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    WebhookEndpointCreate:
      type: object
      required: [url, secret]
      properties:
        url:     { type: string, format: uri }
        secret:  { type: string, minLength: 1, description: "HMAC-SHA256 signing key" }
        events:
          type: array
          items: { $ref: "#/components/schemas/WebhookEventType" }
        format:  { $ref: "#/components/schemas/WebhookFormat" }
        enabled: { type: boolean, default: true }
    WebhookEndpointUpdate:
      type: object
      required: [url, secret]
      description: |
        Full replacement — caller must supply current values for fields that should
        not change (URL, secret, events, format, enabled).
      properties:
        url:     { type: string, format: uri }
        secret:  { type: string, minLength: 1 }
        events:
          type: array
          items: { $ref: "#/components/schemas/WebhookEventType" }
        format:  { $ref: "#/components/schemas/WebhookFormat" }
        enabled: { type: boolean }
    ListWebhookEndpointsResponse:
      type: object
      required: [endpoints]
      properties:
        endpoints:
          type: array
          items: { $ref: "#/components/schemas/WebhookEndpoint" }
...